+++ src/components/TodoList.js
@@ -1,42 +1,42 @@
-import React, {useState} from 'react';
-import { Todolist, Form } from './todo'
+import React, {useEffect, useState} from 'react';
+import { Todolist } from './todo';
 
 const TodoList = () => {
 	const [state, setState] = useState({
-		nextId: 3,
-		todos: [
-			{
-			  id: 0,
-			  todoText: 'Make a todo app',
-			  completed: false
-			}, {
-			  id: 1,
-			  todoText: 'Put in a Rhea easter egg',
-			  completed: false
-			}, {
-			  id: 2,
-			  todoText: 'Profit',
-			  completed: false
-			},
-		  ]
-
+		nextId: 0,
+		todos: []
 	})
 
+	useEffect(() => {
+		const json = localStorage.getItem("todos");
+		const laodTodo = JSON.parse(json);
+		if(laodTodo) {
+			setState(laodTodo);
+		}
+	}, [])
+	useEffect(() => {
+		const json = JSON.stringify(state);
+		localStorage.setItem("todos", json)
+	}, [state]);
+
 	const createTodo = newTodo => {
-		console.log(newTodo);
-		let todo = {id: [state.nextId], todoText: newTodo, completed: false};
+		let todo = {
+			id: [state.nextId],
+			todoText: newTodo,
+			completed: false
+		};
 		let todos = [...state.todos, todo];
 		let nextId = state.nextId + 1;
 		setState({todos, nextId});
-	  }
-  
+	}
+
 	const deleteTodo = id => {
 		let todos = state.todos.filter(todo => (
 		  todo.id !== id
 		));
 		setState({todos});
-	  }
-  
+	}
+
 	const  toggleTodoCompletion = id => {
 		let todos = state.todos.map(todo => (
 		  (todo.id === id)
@@ -44,8 +44,8 @@ const TodoList = () => {
 		  : todo
 		));
 		setState({todos});
-	  }
-	  
+	}
+ 
 	const  editTodo = (id, newText) => {
 		let todos = state.todos.map(todo => (
 		  (todo.id === id)
@@ -53,8 +53,29 @@ const TodoList = () => {
 		  : todo
 		));
 		setState({todos});
-	  }
-	  
+	}
+
+	function Form({createTodo}) {
+		function handleSubmit(e) {
+			e.preventDefault();
+			createTodo(e.target.inputText.value);
+			e.target.inputText.value = '';
+		}
+		
+		return (
+			<form 
+			onSubmit={handleSubmit}
+			>
+			<input 
+				type="text"
+				className="inputText"
+				name="inputText"
+				placeholder="add todo"
+			/>
+			</form>
+		);
+	}
+
 		return (
 		  <div className="todolist">
 			<Form createTodo={createTodo} />
